Mise à jour des propriétés physiques des classes Balle et Robot; amélioration de la gestion de la position et des collisions dans Entite et Position2D.

+++ src/main/java/mkx/imtminesales/robot2d/core/Position2D.java
@@ -7,43 +7,43 @@ package mkx.imtminesales.robot2d.core;
  */
 public class Position2D {
 
-    private int x;
-    private int y;
-    private int lastX;
-    private int lastY;
+    private double x;
+    private double y;
+    private double lastX;
+    private double lastY;
 
-    public Position2D(int x, int y) {
+    public Position2D(double x, double y) {
         this.x = x;
         this.y = y;
         this.lastX = x;
         this.lastY = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getLastX() {
+    public double getLastX() {
         return lastX;
     }
 
-    public int getLastY() {
+    public double getLastY() {
         return lastY;
     }
 
-    public void deplacer(int dx, int dy) {
+    public void deplacer(double dx, double dy) {
         lastX = x;
         lastY = y;
 
@@ -57,8 +57,8 @@ public class Position2D {
     }
 
     public double calculerDistance(Position2D autre) {
-        int dx = x - autre.x;
-        int dy = y - autre.y;
+        double dx = x - autre.x;
+        double dy = y - autre.y;
         return Math.sqrt(dx * dx + dy * dy);
     }
 }
